+++ lib/grape/validations/coerce.rb
@@ -46,11 +46,11 @@ module Grape
       end
 
       def coerce_value(type, val)
-        converter = Virtus::Attribute.build(:a, type)
+        converter = Virtus::Attribute.build(type)
         converter.coerce(val)
 
       # not the prettiest but some invalid coercion can currently trigger
-      # errors in Virtus (see coerce_spec.rb)
+      # errors in Virtus (see coerce_spec.rb:75)
       rescue
         InvalidValue.new
       end